CssFn replacing CssFnResult, getValue util, type naming nits for consistency, updated snapshots
