+++ .eslintrc.yml
@@ -1,0 +1,8 @@
+extends:
+  - "@postmates"
+
+settings:
+  import/resolver:
+    node:
+      extensions:
+        - ".ts"
