+++ .gitignore
@@ -4,7 +4,6 @@ node_modules
 
 # build
 dist
-.rpt2_cache
 
 # test
 .nyc_output
