--- .npmignore
@@ -1,1 +1,0 @@
-dist/gen-types
