--- bench/README.md
@@ -1,26 +1,0 @@
-
-# Benchmarks
-
-## v0.1.5
-```
-generateFonts x 205,608 ops/sec ±0.87% (92 runs sampled)
-generateFontFace x 197,131 ops/sec ±0.21% (93 runs sampled)
-```
-
-## v0.1.4
-```
-generateFonts x 202,982 ops/sec ±0.93% (88 runs sampled)
-generateFontFace x 208,720 ops/sec ±0.26% (94 runs sampled)
-```
-
-## v0.1.3
-```
-generateFonts x 18,139 ops/sec ±0.61% (90 runs sampled)
-generateFontFace x 227,436 ops/sec ±0.30% (95 runs sampled)
-```
-
-## v0.1.2
-```
-generateFonts x 15,107 ops/sec ±0.87% (90 runs sampled)
-generateFontFace x 202,995 ops/sec ±0.35% (93 runs sampled)
-```
