+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.11",
   "description": "A quest for consistency",
   "main": "./dist/index.js",
-  "module": "./dist/index.es.js",
   "types": "./dist/index.d.ts",
   "sideEffects": false,
   "files": [
@@ -14,15 +13,10 @@
   },
   "scripts": {
     "prebuild": "rm -rf dist",
-    "postbuild": "yarn run types",
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "types": "yarn dts-bundle-generator -o ./dist/index.d.ts ./dist/gen-types/index.d.ts",
-    "lint": "tslint src/**/*.ts --project tsconfig.json",
+    "build": "tsc",
     "test": "nyc ava",
-    "test:size": "bundlesize",
-    "test:codecov": "nyc report --reporter=lcov > coverage.lcov && npx codecov",
-    "bench": "node bench",
+    "lint": "yarn run eslint --ext .ts src",
+    "coverage": "nyc report --reporter=lcov > coverage.lcov && npx codecov",
     "prepare": "yarn run -s build"
   },
   "dependencies": {
@@ -30,25 +24,24 @@
   },
   "devDependencies": {
     "@emotion/core": "^10.0.10",
-    "@types/node": "^11.11.6",
+    "@postmates/eslint-config": "^4.0.1",
+    "@types/node": "^11.13.4",
+    "@types/react": "^16.8.13",
     "@types/rollup-plugin-node-resolve": "^4.0.0",
-    "ava": "^1.3.1",
-    "benchmark": "^2.1.4",
-    "bundlesize": "^0.17.1",
-    "dts-bundle-generator": "^2.1.0",
-    "esm": "^3.2.18",
+    "@typescript-eslint/eslint-plugin": "^1.6.0",
+    "@typescript-eslint/parser": "^1.6.0",
+    "ava": "^1.4.1",
+    "eslint": "^5.16.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.12.4",
+    "esm": "^3.2.22",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.5",
     "nyc": "^13.3.0",
     "react": "^16.8.6",
-    "rollup": "^1.6.0",
-    "rollup-plugin-node-resolve": "^4.0.1",
-    "rollup-plugin-terser": "^4.0.4",
-    "rollup-plugin-typescript2": "^0.20.1",
     "ts-node": "^8.0.3",
-    "tslint": "^5.14.0",
-    "tslint-config-airbnb": "^5.11.1",
-    "typescript": "^3.3.3333"
+    "typescript": "^3.4.3"
   },
   "ava": {
     "verbose": true,
@@ -63,18 +56,14 @@
   },
   "nyc": {
     "all": true,
-    "include": [
-      "src/*.ts"
-    ],
-    "exclude": [
-      "dist",
-      "test"
-    ],
     "extension": [
       ".ts"
     ],
-    "require": [
-      "ts-node/register"
+    "exclude": [
+      "**/*.d.ts",
+      "node_modules/**",
+      "coverage/**",
+      "test/**"
     ],
     "reporter": [
       "lcov",
@@ -83,7 +72,7 @@
   },
   "lint-staged": {
     "*.ts": [
-      "tslint --fix",
+      "eslint --fix",
       "git add"
     ]
   },
@@ -93,16 +82,6 @@
       "pre-commit": "lint-staged"
     }
   },
-  "bundlesize": [
-    {
-      "path": "dist/index.js",
-      "maxSize": "1 kb"
-    },
-    {
-      "path": "dist/index.es.js",
-      "maxSize": "1 kb"
-    }
-  ],
   "homepage": "https://github.com/buames/typesettings-js#readme",
   "repository": "https://github.com/buames/typesettings-js",
   "author": "buames",
