+++ test/snapshots/index.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/index.ts`
+# Snapshot report for `test/index.test.ts`
 
-The actual snapshot is saved in `index.ts.snap`.
+The actual snapshot is saved in `index.test.ts.snap`.
 
 Generated by [AVA](https://ava.li).
