--- tslint.yml
@@ -1,12 +1,0 @@
-extends:
-  - 'tslint-config-airbnb'
-
-rules:
-  indent:
-    - error
-    - 2
-  trailing-comma:
-    - true
-    - multiline: never
-      singleline: never
-  import-name: false
